Proposal: use Type[] in PHPDoc block
To improve support of the eclipse PDT pluggin (for autocompletion), I propose to change the array notation in PHPDoc blocks to match the phpDocumentor notation for "array of type".

See http://www.phpdoc.org/docs/latest/for-users/types.html?highlight=array

See https://github.com/eclipse/pdt/pull/6

For the moment I just modified this file to show the proposal. If it's accepted, we must to change all the PHPDoc in only one PR. It could be done easily with a find and replace.

What do you think about this?